+++ GetandTake/Business/Services/Concrete/SupplierManager.cs
@@ -1,22 +1,22 @@
-﻿using GetandTake.DataAccess.Repositories.Abstract;
+﻿using GetandTake.Business.Services.Abstract;
+using GetandTake.DataAccess.Repositories.Abstract;
 using GetandTake.Models;
-using GetandTake.Services.Abstract;
 
-namespace GetandTake.Services.Concrete;
+namespace GetandTake.Business.Services.Concrete;
 
 public class SupplierManager : ISupplierService
 {
     private readonly ISupplierRepository _repository;
 
-    public SupplierManager(ISupplierRepository repository) => 
+    public SupplierManager(ISupplierRepository repository) =>
         _repository = repository;
 
-    public async Task CreateAsync(Supplier supplier) => 
+    public async Task CreateAsync(Supplier supplier) =>
         await _repository.CreateAsync(supplier);
- 
-    public void Delete(int supplierId) => 
+
+    public void Delete(int supplierId) =>
         _repository.Delete(entity => entity.SupplierID == supplierId);
- 
+
     public async Task UpdateAsync(int supplierId, Supplier supplier)
     {
         var findSupplier = await _repository.GetAsync(category => category.SupplierID == supplierId);
@@ -27,9 +27,9 @@ public class SupplierManager : ISupplierService
         }
     }
 
-    public async Task<List<Supplier>> GetAllAsync() => 
+    public async Task<List<Supplier>> GetAllAsync() =>
         await _repository.GetItemsAsync();
 
-    public async Task<Supplier> GetByIdAsync(int supplierId) => 
+    public async Task<Supplier> GetByIdAsync(int supplierId) =>
         await _repository.GetAsync(supplier => supplier.SupplierID == supplierId);
 }
