+++ GetandTake/Core/Extensions/ImageCachingMiddlewareExtensions.cs
@@ -1,4 +1,4 @@
-﻿using GetandTake.Core.Middlewares;
+﻿using GetandTake.Middlewares;
 
 namespace GetandTake.Core.Extensions;
 
@@ -7,6 +7,6 @@ public static class ImageCachingMiddlewareExtensions
     public static IApplicationBuilder UseMyCustomMiddleware(
         this IApplicationBuilder builder)
     {
-        return builder.UseMiddleware<ImageCachingMiddleware>();
+        return builder.UseMiddleware<CachingMiddleware>();
     }
 }
