+++ GetandTake/Core/Utilities/Interceptors/MethodInterception.cs
@@ -1,12 +1,12 @@
 ﻿using Castle.DynamicProxy;
 
-namespace GetandTake.Core.Interceptors;
+namespace GetandTake.Core.Utilities.Interceptors;
 
 public abstract class MethodInterception : MethodInterceptionBaseAttribute
 {
     protected virtual void OnBefore(IInvocation invocation) { }
     protected virtual void OnAfter(IInvocation invocation) { }
-    protected virtual void OnException(IInvocation invocation, System.Exception e) { }
+    protected virtual void OnException(IInvocation invocation, Exception e) { }
     protected virtual void OnSuccess(IInvocation invocation) { }
     public override void Intercept(IInvocation invocation)
     {
