+++ GetandTake/Core/Utilities/Interceptors/MethodInterceptionBaseAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using Castle.DynamicProxy;
 
-namespace GetandTake.Core.Interceptors;
+namespace GetandTake.Core.Utilities.Interceptors;
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = true, Inherited = true)]
 public abstract class MethodInterceptionBaseAttribute : Attribute, IInterceptor
