+++ GetandTake/Pages/Categories/Image.cshtml.cs
@@ -1,5 +1,5 @@
+using GetandTake.Business.Services.Abstract;
 using GetandTake.Models;
-using GetandTake.Services.Abstract;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace GetandTake.Pages.Categories
